libs/cli: clean up package
